requirements.txt
# general  modules
django >3.2,<3.3
python-decouple
Pillow
djangorestframework>3.12,<3.13
markdown       
django-filter 
django-cors-headers
drf-yasg[validation]
djangorestframework-simplejwt
djoser
django-templated-mail


flake8
black